+++ inst/shinyApp/www/js/customMessageHandlers.js
@@ -3,6 +3,7 @@ var actionRegistration_minifier = null,
   actionRegistration_sass = null,
   actionRegistration_clangFormat = null,
   actionRegistration_cppCheck = null,
+  actionRegistration_brittany = null,
   actionRegistration_styler = null,
   actionRegistration_formatR = null,
   actionRegistration_svgChecker = null,
@@ -11,7 +12,9 @@ var actionRegistration_minifier = null,
   actionRegistration_wordWrapper = null,
   actionRegistration_typescript = null,
   actionRegistration_formatCodeApi = null,
-  actionRegistration_markdownit = null;
+  actionRegistration_markdownit = null,
+  actionRegistration_xmllint = null,
+  actionRegistration_xml2 = null;
 
 function prettifier(parser, label) {
   if(typeof label === "undefined") {
@@ -99,6 +102,9 @@ function formatCodeApi(language, label) {
       break;
     case "swift":
       url = "http://www.zafuswitchout.com:3001/swift";
+      break;
+    case "xml":
+      url = "http://aozozo.com:600/xml";
       break;
   }
   return {
@@ -196,6 +202,9 @@ function actionRegistration(language) {
   if(actionRegistration_cppCheck !== null) {
     actionRegistration_cppCheck.dispose();
   }
+  if(actionRegistration_brittany !== null) {
+    actionRegistration_brittany.dispose();
+  }
   if(actionRegistration_styler !== null) {
     actionRegistration_styler.dispose();
   }
@@ -222,6 +231,12 @@ function actionRegistration(language) {
   }
   if(actionRegistration_markdownit !== null) {
     actionRegistration_markdownit.dispose();
+  }
+  if(actionRegistration_xmllint !== null) {
+    actionRegistration_xmllint.dispose();
+  }
+  if(actionRegistration_xml2 !== null) {
+    actionRegistration_xml2.dispose();
   }
   var languages_formatCodeApi = [
     "apex",
@@ -415,7 +430,7 @@ function actionRegistration(language) {
       contextMenuGroupId: "navigation",
       contextMenuOrder: 1.5,
       run: function(ed) {
-        if(clangFormat) {
+        if(executables.clangFormat) {
           var bookmark = $("#bookmark").prop("checked");
           if(bookmark) {
             var modelId = ed.getModel().id;
@@ -456,7 +471,7 @@ function actionRegistration(language) {
         contextMenuGroupId: "navigation",
         contextMenuOrder: 1.5,
         run: function(ed) {
-          if(cppCheck) {
+          if(executables.cppCheck) {
             var fileName =
               $(chromeTabs.activeTabEl).find(".chrome-tab-title").html();
             Shiny.setInputValue("cppCheck", {
@@ -489,6 +504,44 @@ function actionRegistration(language) {
           formatCodeApi("java", "Prettify (formatCodeApi)")
         );
     }
+  } else if(language === "haskell") { /*                              haskell */
+    actionRegistration_brittany = editor.addAction({
+      id: "brittany",
+      label: "Prettify",
+      precondition: null,
+      keybindingContext: null,
+      contextMenuGroupId: "navigation",
+      contextMenuOrder: 1.5,
+      run: function(ed) {
+        if(executables.brittany) {
+          var bookmark = $("#bookmark").prop("checked");
+          if(bookmark) {
+            var modelId = ed.getModel().id;
+            modelValues[modelId] = ed.getValue();
+            $(chromeTabs.activeTabEl)
+              .find(".chrome-tab-title")
+                .css("font-style", "normal");
+          }
+          Shiny.setInputValue("brittany", ed.getValue());
+        } else {
+          flashFunction({
+            message: "Either <span style='font-style: monospace;'>brittany</span> is not installed or it is not in the PATH variable.",
+            title: "<span style='font-style: monospace;'>brittany</span> not found",
+            type: "info",
+            icon: "glyphicon glyphicon-ban-circle",
+            withTime: true,
+            autoClose: true,
+            closeTime: 10000,
+            animation: true,
+            animShow: "rotateInDownLeft",
+            animHide: "bounceOutRight",
+            position: ["bottom-left", [0, 0.01]],
+            speed: "slow"
+          });
+        }
+        return null;
+      }
+    });
   } else if(language === "r") { /*                                          r */
     actionRegistration_styler = editor.addAction({
       id: "styler",
@@ -664,9 +717,6 @@ function actionRegistration(language) {
     });
     actionRegistration_prettifier =
       editor.addAction(prettifier("html"));
-  } else if(language === "xml") { /*                                       xml */
-    actionRegistration_prettifier =
-      editor.addAction(prettifier("html"));
   } else if(language === "plaintext" || language === undefined) {/* plaintext */
     actionRegistration_wordWrapper =
       editor.addAction(wordWrapper());
@@ -776,6 +826,66 @@ function actionRegistration(language) {
   } else if(language === "yaml") { /*                                    yaml */
     actionRegistration_prettifier =
       editor.addAction(prettifier("yaml"));
+  } else if(language === "xml") { /*                                      xml */
+    actionRegistration_xmllint = editor.addAction({
+      id: "xmllint",
+      label: "Prettify (xmllint)",
+      precondition: null,
+      keybindingContext: null,
+      contextMenuGroupId: "navigation",
+      contextMenuOrder: 1.5,
+      run: function(ed) {
+        if(executables.xmllint) {
+          var bookmark = $("#bookmark").prop("checked");
+          if(bookmark) {
+            var modelId = ed.getModel().id;
+            modelValues[modelId] = ed.getValue();
+            $(chromeTabs.activeTabEl)
+              .find(".chrome-tab-title")
+                .css("font-style", "normal");
+          }
+          Shiny.setInputValue("xmllint", ed.getValue());
+        } else {
+          flashFunction({
+            message: "Either <span style='font-style: monospace;'>xmllint</span> is not installed or it is not in the PATH variable.",
+            title: "<span style='font-style: monospace;'>xmllint</span> not found",
+            type: "info",
+            icon: "glyphicon glyphicon-ban-circle",
+            withTime: true,
+            autoClose: true,
+            closeTime: 10000,
+            animation: true,
+            animShow: "rotateInDownLeft",
+            animHide: "bounceOutRight",
+            position: ["bottom-left", [0, 0.01]],
+            speed: "slow"
+          });
+        }
+        return null;
+      }
+    });
+    actionRegistration_xml2 = editor.addAction({
+      id: "xml2",
+      label: "Prettify (R)",
+      precondition: null,
+      keybindingContext: null,
+      contextMenuGroupId: "navigation",
+      contextMenuOrder: 1.5,
+      run: function(ed) {
+        var bookmark = $("#bookmark").prop("checked");
+        if(bookmark) {
+          var modelId = ed.getModel().id;
+          modelValues[modelId] = ed.getValue();
+          $(chromeTabs.activeTabEl)
+            .find(".chrome-tab-title")
+              .css("font-style", "normal");
+        }
+        Shiny.setInputValue("xml2", ed.getValue());
+        return null;
+      }
+    });
+    actionRegistration_formatCodeApi =
+      editor.addAction(formatCodeApi("xml", "Prettify (formatCodeApi)"));
   } else if(languages_formatCodeApi.indexOf(language) > -1) {/* formatCodeApi */
     actionRegistration_formatCodeApi =
       editor.addAction(formatCodeApi(language));
@@ -790,8 +900,8 @@ function setModel(valueAndLanguage) {
     language
   );
   modelInstance.updateOptions({
-    tabSize: 2,
-    indentSize: 2
+    tabSize: tabSize,
+    indentSize: tabSize
   });
   modelInstance.onDidChangeContent((event) => {
     $(chromeTabs.activeTabEl)
@@ -814,14 +924,6 @@ function setValue(value) {
   editor.setValue(value);
 }
 
-function setClangFormat(x) {
-  clangFormat = x;
-}
-
-function setCppCheck(x) {
-  cppCheck = x;
-}
-
 function changeBorders(id) {
   $(`label[for=${id}]`).next().find(".form-control")
     .css("border-bottom-right-radius", 0);
@@ -840,8 +942,6 @@ $(document).on("shiny:connected", function() {
   Shiny.addCustomMessageHandler("modelInstance", setModel);
   Shiny.addCustomMessageHandler("language", setLanguage);
   Shiny.addCustomMessageHandler("value", setValue);
-  Shiny.addCustomMessageHandler("clangFormat", setClangFormat);
-  Shiny.addCustomMessageHandler("cppCheck", setCppCheck);
   Shiny.addCustomMessageHandler("flashMessage", flashFunction);
   Shiny.addCustomMessageHandler("changeBorders", changeBorders);
   Shiny.addCustomMessageHandler("setTheme", setTheme);