+++ NEWS.md
@@ -1,6 +1,15 @@
+# shinyMonacoEditor 1.1.0
+
+* Haskell formatter, requires `brittany`.
+* XML formatter requiring `xmllint`.
+* XML formatter using the `xml2` package.
+* Option `indentSize`, to set the number of spaces of the indentation.
+
+
+
 # shinyMonacoEditor 1.0.0
 
-* Updated 'prettier'.
+* Updated `prettier`.
 * A trailing comma was added in JavaScript objects when prettifying JavaScript or HTML; this does not happen anymore.
 * New feature: SVG rescaling. 
 * New feature: SVG panning and zooming.